add a new types folder for organizing logTypes
