+++ packages/logging-controller/src/index.ts
@@ -1,1 +1,2 @@
 export * from './LoggingController';
+export * from './logTypes';
